--- infrastructure/aws-config.ts
@@ -1,4 +1,0 @@
-export const awsConfig = {
-  profile: "brianpatrickkemper-sandbox",
-  region: "us-east-1",
-};
