+++ pnpm-workspace.yaml
@@ -1,0 +1,2 @@
+packages:
+  - "web"
