+++ web/app/favicon.ico
