+++ web/components/logos/SparkpostLogo.tsx
