+++ web/components/logos/StaqLogo.tsx
