+++ web/components/logos/VisualLeaseLogo.tsx
