+++ web/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@bpk/web",
+  "description": "You found me!",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "build": "next build",
+    "dev": "next dev",
+    "favicon": "magick -background transparent public/logo.png -define icon:auto-resize=16,24,32,48,64,72,96,128,256 app/favicon.ico",
+    "lint": "next lint",
+    "set:timezone": "sudo systemsetup -settimezone",
+    "start": "next start"
+  },
+  "dependencies": {
+    "@radix-ui/react-dialog": "^1.1.6",
+    "@radix-ui/react-icons": "^1.3.2",
+    "@radix-ui/react-popover": "^1.1.6",
+    "@radix-ui/react-toggle": "^1.1.2",
+    "@radix-ui/react-tooltip": "^1.1.8",
+    "classnames": "^2.5.1",
+    "ms": "^2.1.3",
+    "next": "^15.2.4",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "usehooks-ts": "^3.1.1"
+  },
+  "devDependencies": {
+    "@eslint/eslintrc": "^3.2.0",
+    "@eslint/js": "^9.20.0",
+    "@tailwindcss/postcss": "^4.0.0",
+    "@types/ms": "^0.7.34",
+    "@types/node": "^22.13.2",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "eslint": "^9.21.0",
+    "eslint-config-next": "^15.1.7",
+    "postcss": "^8",
+    "tailwindcss": "^4.0.0",
+    "typescript": "^5.7.3"
+  },
+  "engines": {
+    "node": "^22.10.0"
+  },
+  "packageManager": "pnpm@10.10.0+sha512.d615db246fe70f25dcfea6d8d73dee782ce23e2245e3c4f6f888249fb568149318637dca73c2c5c8ef2a4ca0d5657fb9567188bfab47f566d1ee6ce987815c39"
+}
